+++ experimental/packages/web-common/src/index.ts
@@ -16,3 +16,25 @@ export type { SessionIdGenerator } from './types/SessionIdGenerator';
 export type { SessionPublisher } from './types/SessionPublisher';
 export type { SessionObserver } from './types/SessionObserver';
 export type { SessionStore } from './types/SessionStore';
+
+export type {
+  PerformanceEntries,
+  PerformanceLegacy,
+  PerformanceResourceTimingInfo,
+} from './types/Performance';
+export type {
+  PropagateTraceHeaderCorsUrls,
+  URLLike,
+} from './utils-resource-timings';
+export { PerformanceTimingNames } from './enums/PerformanceTimingNames';
+export {
+  addSpanNetworkEvent,
+  addSpanNetworkEvents,
+  getElementXPath,
+  getResource,
+  hasKey,
+  normalizeUrl,
+  parseUrl,
+  shouldPropagateTraceHeaders,
+  sortResources,
+} from './utils-resource-timings';